Muttythreading calcualte frame, particle effect

+++ TacticsVictory/common_files/Network/Messages/MessagesServer_.cpp
@@ -33,18 +33,18 @@ bool MessageCreateLandscape::Decompress(Decompressor &data)
 void MessageCreateGameObject::Compress(Compressor &data) const
 {
     data << num_objects;
-    data.Write(id, sizeof(id[0]) * num_objects);
-    data.Write(type, sizeof(type[0][0]) * num_objects * 3);
-    data.Write(transform, sizeof(transform[0]) * num_objects);
+    data.Write(id,        (uint)sizeof(id[0]) * num_objects);
+    data.Write(type,      (uint)sizeof(type[0][0]) * num_objects * 3);
+    data.Write(transform, (uint)sizeof(transform[0]) * num_objects);
 }
 
 
 bool MessageCreateGameObject::Decompress(Decompressor &data)
 {
     data >> num_objects;
-    data.Read(id, sizeof(id[0]) * num_objects);
-    data.Read(type, sizeof(type[0][0]) * num_objects * 3);
-    data.Read(transform, sizeof(transform[0]) * num_objects);
+    data.Read(id,        (uint)sizeof(id[0]) * num_objects);
+    data.Read(type,      (uint)sizeof(type[0][0]) * num_objects * 3);
+    data.Read(transform, (uint)sizeof(transform[0]) * num_objects);
 
     return true;
 }
@@ -54,13 +54,10 @@ void MessageGameObjectState::Compress(Compressor &data) const
 {
     data << num_objects;
 
-    data.Write(id, sizeof(id[0]) * num_objects);
-
-    data.Write(position, sizeof(position[0]) * num_objects);
-
-    data.Write(direction, sizeof(direction[0]) * num_objects);
-
-    data.Write(up, sizeof(up[0]) * num_objects);
+    data.Write(id,        (uint)sizeof(id[0]) * num_objects);
+    data.Write(position,  (uint)sizeof(position[0]) * num_objects);
+    data.Write(direction, (uint)sizeof(direction[0]) * num_objects);
+    data.Write(up,        (uint)sizeof(up[0]) * num_objects);
 }
 
 
@@ -68,13 +65,10 @@ bool MessageGameObjectState::Decompress(Decompressor &data)
 {
     data >> num_objects;
 
-    data.Read(id, sizeof(id[0]) * num_objects);
-
-    data.Read(position, sizeof(position[0]) * num_objects);
-
-    data.Read(direction, sizeof(direction[0]) * num_objects);
-
-    data.Read(up, sizeof(up[0]) * num_objects);
+    data.Read(id,        (uint)sizeof(id[0]) * num_objects);
+    data.Read(position,  (uint)sizeof(position[0]) * num_objects);
+    data.Read(direction, (uint)sizeof(direction[0]) * num_objects);
+    data.Read(up,        (uint)sizeof(up[0]) * num_objects);
 
     return true;
 }
